1. Исправлена ошибка, из-за котой неправильно расчитывалась скорость 2. Рефакторинг java кода
